# cuda4dnn(mish): improve accuracy and performance